requirements.txt
Pillow>=10.0.0
numpy>=1.24.0
PyYAML>=6.0
flask>=3.0.0
